# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SIAKAD GERBANG",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reference/grade-lists": {
            "get": {
                "tags": [
                    "Reference/GradeList"
                ],
                "summary": "List all GradeList",
                "operationId": "6d20dacc2b94fc460d93aa0411bde05c",
                "responses": {
                    "200": {
                        "description": "List of GradeList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeListListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference/GradeList"
                ],
                "summary": "Create a new GradeList",
                "operationId": "0013fb31b71cacfc80d16a10708bc503",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "GradeList created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/reference/grade-lists/{id}": {
            "get": {
                "tags": [
                    "Reference/GradeList"
                ],
                "summary": "Get GradeList detail by ID",
                "operationId": "1f3bf64f21ac9a5ea3ac75e2ba3d826c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "GradeList ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GradeList detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeListResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "GradeList not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reference/GradeList"
                ],
                "summary": "Update an existing GradeList",
                "operationId": "1c7bb649ebddd7b605ad7802363faf1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "GradeList ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "GradeList updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeListResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "GradeList not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reference/GradeList"
                ],
                "summary": "Delete a GradeList by ID",
                "operationId": "34f155b36e6592e40ea7a6af4060169f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "GradeList ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GradeList deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Data deleted",
                                    "data": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "GradeList not found"
                    }
                }
            }
        },
        "/api/reference/graduations": {
            "get": {
                "tags": [
                    "Reference/Graduation"
                ],
                "summary": "List all Graduation",
                "operationId": "c02fb9f9e0035850afdf8e6dfa0380b1",
                "responses": {
                    "200": {
                        "description": "List of Graduation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GraduationListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference/Graduation"
                ],
                "summary": "Create a new Graduation",
                "operationId": "7265e557a104bd15480c85b89b347cdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Graduation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GraduationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/reference/graduations/{id}": {
            "get": {
                "tags": [
                    "Reference/Graduation"
                ],
                "summary": "Get Graduation detail by ID",
                "operationId": "8691201d364586dbac5b5544f22927c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Graduation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Graduation detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GraduationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Graduation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reference/Graduation"
                ],
                "summary": "Update an existing Graduation",
                "operationId": "a85530b398e22f8d79e03b62a404f90f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Graduation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Graduation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GraduationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Graduation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reference/Graduation"
                ],
                "summary": "Delete a Graduation by ID",
                "operationId": "9ff8f4106a2803b2c5e207b5fbcba587",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Graduation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Graduation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Data deleted",
                                    "data": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Graduation not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GradeList": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "B"
                    },
                    "gpa_score": {
                        "type": "string",
                        "example": "3.00"
                    }
                },
                "type": "object"
            },
            "GradeListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Grade detail"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GradeList"
                    }
                },
                "type": "object"
            },
            "GradeListListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "List of grade lists"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GradeList"
                        }
                    }
                },
                "type": "object"
            },
            "Graduation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "year": {
                        "type": "string",
                        "example": "2025-2026"
                    },
                    "note": {
                        "type": "json",
                        "example": "{}"
                    }
                },
                "type": "object"
            },
            "GraduationResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Graduation detail"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Graduation"
                    }
                },
                "type": "object"
            },
            "GraduationListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "List of graduations"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Graduation"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}